Fixing traffic server logging for new version and adding tests

+++ test/support/log_tail.rb
@@ -1,18 +1,26 @@
 class LogTail
   def initialize(filename)
     @path = File.join($config["log_dir"], filename)
-    File.open(@path) do |file|
-      file.seek(0, IO::SEEK_END)
-      @pos = file.pos
+    if File.exist?(@path)
+      File.open(@path) do |file|
+        file.seek(0, IO::SEEK_END)
+        @pos = file.pos
+      end
+    else
+      @pos = 0
     end
   end
 
   def read
     output = nil
-    File.open(@path) do |file|
-      file.seek(@pos)
-      output = file.read.encode("UTF-8", invalid: :replace)
-      @pos = file.pos
+    if File.exist?(@path)
+      File.open(@path) do |file|
+        file.seek(@pos)
+        output = file.read.encode("UTF-8", invalid: :replace)
+        @pos = file.pos
+      end
+    else
+      output = ""
     end
 
     output
